+++ src/Firebase/Auth/IsTenantAware.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kreait\Firebase\Auth;
+
+interface IsTenantAware
+{
+    public function tenantId(): ?TenantId;
+}