+++ src/Firebase/Auth/SignInWithCustomToken.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Auth;
 
-final class SignInWithCustomToken implements SignIn
+final class SignInWithCustomToken implements IsTenantAware, SignIn
 {
     /** @var string */
     private $customToken;
+
+    /** @var TenantId|null */
+    private $tenantId;
 
     private function __construct()
     {
@@ -21,8 +24,21 @@ final class SignInWithCustomToken implements SignIn
         return $instance;
     }
 
+    public function withTenantId(TenantId $tenantId): self
+    {
+        $action = clone $this;
+        $action->tenantId = $tenantId;
+
+        return $action;
+    }
+
     public function customToken(): string
     {
         return $this->customToken;
+    }
+
+    public function tenantId(): ?TenantId
+    {
+        return $this->tenantId;
     }
 }